pyproject.toml
[project]
name = "cartesapp"
version = "0.0.1"
authors = [
  { name="Lyno Ferraz", email="lyno@prototyp3.dev" },
]
description = "A Python Opinionated High Framework for Cartesi Distributed Applications"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-cartesi >= 0.1.0",
    "Jinja2 >= 3.1.3",
    "pony >= 0.7.17",
    "typer >= 0.9.0",
    "pydantic-to-typescript >= 1.0.10",
]

[project.urls]
"Homepage" = "https://github.com/prototyp3-dev/cartesapp"

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"
